BUGFIX: Reuse existing controller context instead of creating a new one
The ResourceMapper is the one in charge for creating all kinds of URIs,
like the self link of an entity or the relation links.

This change makes the ResourceMapper not create a new ControllerContext
but rely on external data during its rendering time.

The ApiController, in turn, is the one passing the ControllerContext
which is currently active along.