fix: Don't disconnect ApiController::$request and $request by cloning
Although cloning the request object and this replacing it is done in the
Flow RestController::resolveActionMethodName method as well "to prevent
unexpected routing behavior", the Flow ActionController::processRequest()
method always uses its passing request object when calling the action
method, never the one stored in the controller state. So cloning the
request object in our ApiController will not pass updated data back to
the controller.